feat: Rotator 적용
추가 디버깅 필요 상태
